STAC-24013: Minor fixes for when global.backup is disabled

+++ internal/scripts/scripts/restore-settings-backup.sh
@@ -6,14 +6,16 @@ export TMP_DIR=/tmp-data
 
 RESTORE_FILE="${BACKUP_DIR}/${BACKUP_FILE}"
 
-export AWS_ACCESS_KEY_ID
-AWS_ACCESS_KEY_ID="$(cat /aws-keys/accesskey)"
-export AWS_SECRET_ACCESS_KEY
-AWS_SECRET_ACCESS_KEY="$(cat /aws-keys/secretkey)"
+if [ "$BACKUP_CONFIGURATION_UPLOAD_REMOTE" == "true" ] && [ ! -f "${RESTORE_FILE}" ]; then
+  export AWS_ACCESS_KEY_ID
+  AWS_ACCESS_KEY_ID="$(cat /aws-keys/accesskey)"
+  export AWS_SECRET_ACCESS_KEY
+  AWS_SECRET_ACCESS_KEY="$(cat /aws-keys/secretkey)"
 
-echo "=== Downloading Settings backup \"${BACKUP_FILE}\" from bucket \"${BACKUP_CONFIGURATION_BUCKET_NAME}\"..."
-sts-toolbox aws s3 --endpoint "http://${MINIO_ENDPOINT}" --region minio cp "s3://${BACKUP_CONFIGURATION_BUCKET_NAME}/${BACKUP_CONFIGURATION_S3_PREFIX}${BACKUP_FILE}" "${TMP_DIR}/${BACKUP_FILE}"
-RESTORE_FILE="${TMP_DIR}/${BACKUP_FILE}"
+  echo "=== Downloading Settings backup \"${BACKUP_FILE}\" from bucket \"${BACKUP_CONFIGURATION_BUCKET_NAME}\"..."
+  sts-toolbox aws s3 --endpoint "http://${MINIO_ENDPOINT}" --region minio cp "s3://${BACKUP_CONFIGURATION_BUCKET_NAME}/${BACKUP_CONFIGURATION_S3_PREFIX}${BACKUP_FILE}" "${TMP_DIR}/${BACKUP_FILE}"
+  RESTORE_FILE="${TMP_DIR}/${BACKUP_FILE}"
+fi
 
 if [ ! -f "${RESTORE_FILE}" ]; then
 echo "=== Backup file \"${RESTORE_FILE}\" not found, exiting..."
